Add recursive job split example.
This is a properly-recursive split where we decide to recurse based on
the size of the input.

This example is terribly far from a real use case. However, I think this
example shows well that there is a pattern of looking at the input, then
splitting the input or not depending on the size.